# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the PPSKSY-Cluster Webapplication",
        "title": "PPSKSY-Cluster API",
        "contact": {},
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/cresources/{id}": {
            "get": {
                "description": "Get cluster resource by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster-resources"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "CResource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CResource"
                        }
                    }
                }
            },
            "put": {
                "description": "Update cluster resource",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cluster-resources"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "CResource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.CResource"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete cluster resource",
                "tags": [
                    "cluster-resources"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "CResource ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CResource": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                }
            }
        },
...
    }
}